Move findEntity and findForeignKeyInEntity into utils file
